% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Hello}
\alias{Hello}
\title{hello
A function to say hello}
\usage{
Hello(name)
}
\arguments{
\item{name}{Name of individual(s) to greet}
}
\value{
A character vector of greetings
}
\description{
hello
A function to say hello
}
